+++ lapdog/vue/src/main.js
@@ -5,6 +5,7 @@ import Home from './Pages/Home.vue'
 import Workspace from './Pages/Workspace.vue'
 import Submission from './Pages/Submission.vue'
 import Methods from './Pages/Methods.vue'
+import Pagination from './Components/Pagination.vue'
 // import Slideout from 'vue-slideout'
 window.$ = require('jquery')
 window.jQuery = require('jquery')
@@ -52,6 +53,8 @@ const router = new Router({
 //     console.log("MOUNTED")
 //   }
 // })
+
+Vue.component('pagination', Pagination);
 
 new Vue({
   el: '#app',